requirements.txt
kedro==0.18.1
openai
tenacity
numpy
python-dotenv
rich